fix type warning

+++ src/ray/gcs/gcs_server/gcs_placement_group_scheduler.cc
@@ -93,9 +93,9 @@ void GcsPlacementGroupScheduler::Schedule(
           auto decision = GetDecision();
           if (status.ok()) {
             decision[pos] = schedule_bundles[pos];
-            for (int i = 0; i < reply.resource_mapping_size(); i++) {
+            for (size_t i = 0; i < reply.resource_mapping_size(); i++) {
               std::string name = std::string(reply.resource_mapping(i).name());
-              for (int j = 0; j < reply.resource_mapping(i).resource_ids_size(); j++) {
+              for (size_t j = 0; j < reply.resource_mapping(i).resource_ids_size(); j++) {
                 int64_t index =
                     int64_t(reply.resource_mapping(i).resource_ids(j).index());
                 double quantity =
@@ -123,7 +123,7 @@ void GcsPlacementGroupScheduler::Schedule(
                     bundle_id, Data, [](Status status) {}));
 
                 rpc::BundleResourceMap resource_map;
-                for (int j = 0; j < resource_lease_[i].size(); j++) {
+                for (size_t j = 0; j < resource_lease_[i].size(); j++) {
                   auto resource = resource_map.add_resource();
                   resource->set_name(std::get<0>(resource_lease_[i][j]));
                   resource->set_index(std::get<1>(resource_lease_[i][j]));
@@ -135,7 +135,7 @@ void GcsPlacementGroupScheduler::Schedule(
               }
               schedule_success_handler_(placement_group);
             } else {
-              for (int i = 0; i < finish_count; i++) {
+              for (size_t i = 0; i < finish_count; i++) {
                 if (decision[i] != ClientID::Nil()) {
                   RetureReourceForNode(bundles[i],
                                        gcs_node_manager_.GetNode(schedule_bundles[pos]));
